feat(experiments): show trace slide-over on experiment page
