+++ app/src/pages/experiment/ExperimentCompareTable.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode, useMemo, useState } from "react";
+import React, { ReactNode, startTransition, useMemo, useState } from "react";
 import { graphql, useLazyLoadQuery } from "react-relay";
 import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
 import { useNavigate } from "react-router";
@@ -46,6 +46,8 @@ import { TableEmpty } from "@phoenix/components/table/TableEmpty";
 import { LatencyText } from "@phoenix/components/trace/LatencyText";
 import { assertUnreachable } from "@phoenix/typeUtils";
 
+import { TraceDetails } from "../trace";
+
 import {
   ExperimentCompareTableQuery,
   ExperimentCompareTableQuery$data,
@@ -58,6 +60,21 @@ type ExampleCompareTableProps = {
    * Whether to display the full text of the text fields
    */
   displayFullText: boolean;
+};
+
+type ExperimentInfoMap = Record<
+  string,
+  { name: string; sequenceNumber: number } | undefined
+>;
+
+type TableRow = ExperimentCompareTableQuery$data["comparisons"][number] & {
+  id: string;
+  input: unknown;
+  referenceOutput: unknown;
+  runComparisonMap: Record<
+    string,
+    ExperimentCompareTableQuery$data["comparisons"][number]["runComparisonItems"][number]
+  >;
 };
 
 type ExperimentRun =
@@ -137,19 +154,13 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
   );
   const experimentInfoById = useMemo(() => {
     return (
-      data.dataset?.experiments?.edges.reduce(
-        (acc, edge) => {
-          acc[edge.experiment.id] = { ...edge.experiment };
-          return acc;
-        },
-        {} as Record<
-          string,
-          { name: string; sequenceNumber: number } | undefined
-        >
-      ) || {}
+      data.dataset?.experiments?.edges.reduce((acc, edge) => {
+        acc[edge.experiment.id] = { ...edge.experiment };
+        return acc;
+      }, {} as ExperimentInfoMap) || {}
     );
   }, [data]);
-  const tableData = useMemo(
+  const tableData: TableRow[] = useMemo(
     () =>
       data.comparisons.map((comparison) => {
         const runComparisonMap = comparison.runComparisonItems.reduce(
@@ -172,8 +183,8 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
       }),
     [data]
   );
-  type TableRow = (typeof tableData)[number];
-  const [selectedExample, setSelectedExample] = useState<TableRow | null>();
+
+  const [dialog, setDialog] = useState<ReactNode>(null);
   const baseColumns: ColumnDef<TableRow>[] = [
     {
       header: "input",
@@ -212,8 +223,78 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
         }
         // Only show the first run
         const run = runComparisonItem?.runs[0];
+
+        let traceButton = null;
+        const traceId = run?.trace?.traceId;
+        const projectId = run?.trace?.projectId;
+        if (traceId && projectId) {
+          traceButton = (
+            <TooltipTrigger>
+              <Button
+                variant="default"
+                className="trace-button"
+                size="compact"
+                aria-label="View run trace"
+                icon={<Icon svg={<Icons.Trace />} />}
+                onClick={(e) => {
+                  e.preventDefault();
+                  e.stopPropagation();
+                  startTransition(() => {
+                    setDialog(
+                      <TraceDetailsDialog
+                        traceId={traceId}
+                        projectId={projectId}
+                      />
+                    );
+                  });
+                }}
+              />
+              <Tooltip>View Trace</Tooltip>
+            </TooltipTrigger>
+          );
+        }
         return run ? (
-          <ExperimentRunOutput {...run} displayFullText={displayFullText} />
+          <ExperimentRunOutput
+            {...run}
+            displayFullText={displayFullText}
+            extra={
+              <div
+                css={css`
+                  position: absolute;
+                  top: 0;
+                  right: 0;
+                  display: flex;
+                  flex-direction: column;
+                  gap: var(--ac-global-dimension-static-size-100);
+                `}
+              >
+                <TooltipTrigger>
+                  <Button
+                    variant="default"
+                    className="expand-button"
+                    size="compact"
+                    aria-label="View example run details"
+                    icon={<Icon svg={<Icons.ExpandOutline />} />}
+                    onClick={(e) => {
+                      e.preventDefault();
+                      e.stopPropagation();
+                      startTransition(() => {
+                        setDialog(
+                          <SelectedExampleDialog
+                            selectedExample={row.original}
+                            datasetId={datasetId}
+                            experimentInfoById={experimentInfoById}
+                          />
+                        );
+                      });
+                    }}
+                  />
+                  <Tooltip>View run details</Tooltip>
+                </TooltipTrigger>
+                {traceButton}
+              </div>
+            }
+          />
         ) : (
           <NotRunText />
         );
@@ -250,6 +331,7 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
   // Make sure the table is at least 1280px wide
   const tableTotalSize =
     table.getTotalSize() > 1280 ? table.getTotalSize() + "px" : "100%";
+
   return (
     <div
       css={css`
@@ -299,12 +381,7 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
         ) : (
           <tbody>
             {rows.map((row) => (
-              <tr
-                key={row.id}
-                onClick={() => {
-                  setSelectedExample(row.original);
-                }}
-              >
+              <tr key={row.id}>
                 {row.getVisibleCells().map((cell) => {
                   return (
                     <td
@@ -331,191 +408,10 @@ export function ExperimentCompareTable(props: ExampleCompareTableProps) {
         isDismissable
         type="slideOver"
         onDismiss={() => {
-          setSelectedExample(null);
+          setDialog(null);
         }}
       >
-        {selectedExample ? (
-          <Dialog
-            title={`Comparing Experiments for Example: ${selectedExample.id}`}
-            size="fullscreen"
-            extra={
-              <ExperimentRowActionMenu
-                datasetId={datasetId}
-                exampleId={selectedExample.id}
-              />
-            }
-          >
-            <PanelGroup
-              direction="vertical"
-              autoSaveId="example-compare-panel-group"
-            >
-              <Panel defaultSize={100}>
-                <div
-                  css={css`
-                    overflow-y: auto;
-                    height: 100%;
-                  `}
-                >
-                  <View overflow="hidden" padding="size-200">
-                    <Flex direction="row" gap="size-200" flex="1 1 auto">
-                      <View width="50%">
-                        <Card
-                          title="Input"
-                          {...defaultCardProps}
-                          bodyStyle={{
-                            padding: 0,
-                            maxHeight: "300px",
-                            overflowY: "auto",
-                          }}
-                          extra={
-                            <CopyToClipboardButton
-                              text={JSON.stringify(selectedExample.input)}
-                            />
-                          }
-                        >
-                          <JSONBlock
-                            value={JSON.stringify(
-                              selectedExample.input,
-                              null,
-                              2
-                            )}
-                          />
-                        </Card>
-                      </View>
-                      <View width="50%">
-                        <Card
-                          title="Reference Output"
-                          {...defaultCardProps}
-                          extra={
-                            <CopyToClipboardButton
-                              text={JSON.stringify(selectedExample.input)}
-                            />
-                          }
-                          bodyStyle={{
-                            padding: 0,
-                            maxHeight: "300px",
-                            overflowY: "auto",
-                          }}
-                        >
-                          <JSONBlock
-                            value={JSON.stringify(
-                              selectedExample.referenceOutput,
-                              null,
-                              2
-                            )}
-                          />
-                        </Card>
-                      </View>
-                    </Flex>
-                  </View>
-                </div>
-              </Panel>
-              <PanelResizeHandle css={resizeHandleCSS} />
-              <Panel defaultSize={200}>
-                <Flex direction="column" height="100%">
-                  <View
-                    paddingStart="size-200"
-                    paddingEnd="size-200"
-                    paddingTop="size-100"
-                    paddingBottom="size-100"
-                    borderBottomColor="dark"
-                    borderBottomWidth="thin"
-                    flex="none"
-                  >
-                    <Heading level={2}>Experiments</Heading>
-                  </View>
-                  <div
-                    css={css`
-                      overflow-y: auto;
-                      height: 100%;
-                      padding: var(--ac-global-dimension-static-size-200);
-                    `}
-                  >
-                    <ul
-                      css={css`
-                        display: flex;
-                        flex-direction: column;
-                        gap: var(--ac-global-dimension-static-size-200);
-                      `}
-                    >
-                      {selectedExample.runComparisonItems.map((runItem) => {
-                        const experiment =
-                          experimentInfoById[runItem.experimentId];
-                        return (
-                          <li key={runItem.experimentId}>
-                            <Card
-                              {...defaultCardProps}
-                              title={experiment?.name}
-                              titleExtra={
-                                <SequenceNumberLabel
-                                  sequenceNumber={
-                                    experiment?.sequenceNumber || 0
-                                  }
-                                />
-                              }
-                            >
-                              <ul>
-                                {runItem.runs.map((run, index) => (
-                                  <li key={index}>
-                                    <Flex direction="row">
-                                      <View flex>
-                                        {run.error ? (
-                                          <View padding="size-200">
-                                            <RunError error={run.error} />
-                                          </View>
-                                        ) : (
-                                          <JSONBlock
-                                            value={JSON.stringify(
-                                              run.output,
-                                              null,
-                                              2
-                                            )}
-                                          />
-                                        )}
-                                      </View>
-                                      <ViewSummaryAside width="size-3000">
-                                        <RunLatency
-                                          startTime={run.startTime}
-                                          endTime={run.endTime}
-                                        />
-                                        <ul
-                                          css={css`
-                                            margin-top: var(
-                                              --ac-global-dimension-static-size-100
-                                            );
-                                            display: flex;
-                                            flex-direction: column;
-                                            gap: var(
-                                              --ac-global-dimension-static-size-100
-                                            );
-                                          `}
-                                        >
-                                          {run.annotations?.edges.map(
-                                            (edge) => (
-                                              <li key={edge.annotation.id}>
-                                                <AnnotationLabel
-                                                  annotation={edge.annotation}
-                                                />
-                                              </li>
-                                            )
-                                          )}
-                                        </ul>
-                                      </ViewSummaryAside>
-                                    </Flex>
-                                  </li>
-                                ))}
-                              </ul>
-                            </Card>
-                          </li>
-                        );
-                      })}
-                    </ul>
-                  </div>
-                </Flex>
-              </Panel>
-            </PanelGroup>
-          </Dialog>
-        ) : null}
+        {dialog}
       </DialogContainer>
     </div>
   );
@@ -574,11 +470,17 @@ function ExperimentRowActionMenu(props: {
  * Display the output of an experiment run.
  */
 function ExperimentRunOutput(
-  props: ExperimentRun & { displayFullText: boolean }
+  props: ExperimentRun & { displayFullText: boolean; extra: ReactNode }
 ) {
-  const navigate = useNavigate();
-  const { output, error, startTime, endTime, annotations, displayFullText } =
-    props;
+  const {
+    output,
+    error,
+    startTime,
+    endTime,
+    annotations,
+    displayFullText,
+    extra,
+  } = props;
   if (error) {
     return <RunError error={error} />;
   }
@@ -586,35 +488,13 @@ function ExperimentRunOutput(
     ? annotations.edges.map((edge) => edge.annotation)
     : [];
 
-  const traceId = props.trace?.traceId;
-  const projectId = props.trace?.projectId;
-  const hasTrace = traceId && projectId;
   return (
     <div
       css={css`
         position: relative;
-        .trace-button {
-          position: absolute;
-          top: 0;
-          right: 0;
-        }
       `}
     >
-      {hasTrace ? (
-        <TooltipTrigger>
-          <Button
-            variant="default"
-            className="trace-button"
-            size="compact"
-            aria-label="View run trace"
-            icon={<Icon svg={<Icons.Trace />} />}
-            onClick={() => {
-              navigate(`/projects/${projectId}/traces/${traceId}`);
-            }}
-          />
-          <Tooltip>View Trace</Tooltip>
-        </TooltipTrigger>
-      ) : null}
+      {extra}
       <View paddingEnd="size-500">
         <Flex direction="column" gap="size-100" height="100%">
           <LargeTextWrap>
@@ -699,5 +579,212 @@ function LargeTextWrap({ children }: { children: ReactNode }) {
     >
       {children}
     </div>
+  );
+}
+
+function SelectedExampleDialog({
+  selectedExample,
+  datasetId,
+  experimentInfoById,
+}: {
+  selectedExample: TableRow;
+  datasetId: string;
+  experimentInfoById: ExperimentInfoMap;
+}) {
+  return (
+    <Dialog
+      title={`Comparing Experiments for Example: ${selectedExample.id}`}
+      size="fullscreen"
+      extra={
+        <ExperimentRowActionMenu
+          datasetId={datasetId}
+          exampleId={selectedExample.id}
+        />
+      }
+    >
+      <PanelGroup direction="vertical" autoSaveId="example-compare-panel-group">
+        <Panel defaultSize={100}>
+          <div
+            css={css`
+              overflow-y: auto;
+              height: 100%;
+            `}
+          >
+            <View overflow="hidden" padding="size-200">
+              <Flex direction="row" gap="size-200" flex="1 1 auto">
+                <View width="50%">
+                  <Card
+                    title="Input"
+                    {...defaultCardProps}
+                    bodyStyle={{
+                      padding: 0,
+                      maxHeight: "300px",
+                      overflowY: "auto",
+                    }}
+                    extra={
+                      <CopyToClipboardButton
+                        text={JSON.stringify(selectedExample.input)}
+                      />
+                    }
+                  >
+                    <JSONBlock
+                      value={JSON.stringify(selectedExample.input, null, 2)}
+                    />
+                  </Card>
+                </View>
+                <View width="50%">
+                  <Card
+                    title="Reference Output"
+                    {...defaultCardProps}
+                    extra={
+                      <CopyToClipboardButton
+                        text={JSON.stringify(selectedExample.input)}
+                      />
+                    }
+                    bodyStyle={{
+                      padding: 0,
+                      maxHeight: "300px",
+                      overflowY: "auto",
+                    }}
+                  >
+                    <JSONBlock
+                      value={JSON.stringify(
+                        selectedExample.referenceOutput,
+                        null,
+                        2
+                      )}
+                    />
+                  </Card>
+                </View>
+              </Flex>
+            </View>
+          </div>
+        </Panel>
+        <PanelResizeHandle css={resizeHandleCSS} />
+        <Panel defaultSize={200}>
+          <Flex direction="column" height="100%">
+            <View
+              paddingStart="size-200"
+              paddingEnd="size-200"
+              paddingTop="size-100"
+              paddingBottom="size-100"
+              borderBottomColor="dark"
+              borderBottomWidth="thin"
+              flex="none"
+            >
+              <Heading level={2}>Experiments</Heading>
+            </View>
+            <div
+              css={css`
+                overflow-y: auto;
+                height: 100%;
+                padding: var(--ac-global-dimension-static-size-200);
+              `}
+            >
+              <ul
+                css={css`
+                  display: flex;
+                  flex-direction: column;
+                  gap: var(--ac-global-dimension-static-size-200);
+                `}
+              >
+                {selectedExample.runComparisonItems.map((runItem) => {
+                  const experiment = experimentInfoById[runItem.experimentId];
+                  return (
+                    <li key={runItem.experimentId}>
+                      <Card
+                        {...defaultCardProps}
+                        title={experiment?.name}
+                        titleExtra={
+                          <SequenceNumberLabel
+                            sequenceNumber={experiment?.sequenceNumber || 0}
+                          />
+                        }
+                      >
+                        <ul>
+                          {runItem.runs.map((run, index) => (
+                            <li key={index}>
+                              <Flex direction="row">
+                                <View flex>
+                                  {run.error ? (
+                                    <View padding="size-200">
+                                      <RunError error={run.error} />
+                                    </View>
+                                  ) : (
+                                    <JSONBlock
+                                      value={JSON.stringify(
+                                        run.output,
+                                        null,
+                                        2
+                                      )}
+                                    />
+                                  )}
+                                </View>
+                                <ViewSummaryAside width="size-3000">
+                                  <RunLatency
+                                    startTime={run.startTime}
+                                    endTime={run.endTime}
+                                  />
+                                  <ul
+                                    css={css`
+                                      margin-top: var(
+                                        --ac-global-dimension-static-size-100
+                                      );
+                                      display: flex;
+                                      flex-direction: column;
+                                      gap: var(
+                                        --ac-global-dimension-static-size-100
+                                      );
+                                    `}
+                                  >
+                                    {run.annotations?.edges.map((edge) => (
+                                      <li key={edge.annotation.id}>
+                                        <AnnotationLabel
+                                          annotation={edge.annotation}
+                                        />
+                                      </li>
+                                    ))}
+                                  </ul>
+                                </ViewSummaryAside>
+                              </Flex>
+                            </li>
+                          ))}
+                        </ul>
+                      </Card>
+                    </li>
+                  );
+                })}
+              </ul>
+            </div>
+          </Flex>
+        </Panel>
+      </PanelGroup>
+    </Dialog>
+  );
+}
+
+function TraceDetailsDialog({
+  traceId,
+  projectId,
+}: {
+  traceId: string;
+  projectId: string;
+}) {
+  const navigate = useNavigate();
+  return (
+    <Dialog
+      title="Run Trace Details"
+      size="fullscreen"
+      extra={
+        <Button
+          variant="default"
+          onClick={() => navigate(`/projects/${projectId}/traces/${traceId}`)}
+        >
+          View Trace in Project
+        </Button>
+      }
+    >
+      <TraceDetails traceId={traceId} projectId={projectId} />
+    </Dialog>
   );
 }
