+++ app/src/pages/trace/index.tsx
@@ -1,1 +1,2 @@
+export * from "./TraceDetails";
 export * from "./TracePage";
